[MAINT] move glm 'public' utils outside of _utils
