+++ nilearn/interfaces/bids/glm.py
@@ -10,9 +10,9 @@ from pathlib import Path
 from nilearn import __version__
 from nilearn._utils import logger
 from nilearn._utils.docs import fill_doc
-from nilearn._utils.glm import coerce_to_dict, make_stat_maps
 from nilearn._utils.helpers import is_matplotlib_installed
 from nilearn._utils.logger import find_stack_level
+from nilearn.glm.utils import coerce_to_dict, make_stat_maps
 from nilearn.surface import SurfaceImage
